fix 3

+++ modules/task_2/napylov_e_contrast/main.cpp
@@ -8,7 +8,7 @@ TEST(Linear_stretch_contrast_OMP, Manual_Calc_1) {
     VecImage image = {0, 1, 1, 1, 0, 1 };
     VecImage correct = { 0, 255, 255, 255, 0, 255 };
     VecImage res = add_contrast_omp(image);
-    for (int i = 0; i < image.size(); i++) {
+    for (size_t i = 0; i < image.size(); i++) {
         ASSERT_EQ(correct[i], res[i]);
     }
 }
@@ -17,7 +17,7 @@ TEST(Linear_stretch_contrast_OMP, Manual_Calc_2) {
     VecImage image = { 5, 123, 28, 240 };
     VecImage correct = { 0, 128, 25, 255 };
     VecImage res = add_contrast_omp(image);
-    for (int i = 0; i < image.size(); i++) {
+    for (size_t i = 0; i < image.size(); i++) {
         ASSERT_EQ(correct[i], res[i]);
     }
 }
@@ -26,7 +26,7 @@ TEST(Linear_stretch_contrast_OMP, Manual_Calc_3) {
     VecImage image = { 57, 120, 99, 32, 17, 64, 1, 55, 200 };
     VecImage correct = { 72, 152, 126, 40, 21, 81, 0, 69, 255 };
     VecImage res = add_contrast_omp(image);
-    for (int i = 0; i < image.size(); i++) {
+    for (size_t i = 0; i < image.size(); i++) {
         ASSERT_EQ(correct[i], res[i]);
     }
 }
@@ -35,7 +35,7 @@ TEST(Linear_stretch_contrast_OMP, Manual_Calc_4) {
     VecImage image = { 1, 2, 3, 4, 5, 1 };
     VecImage correct = { 0, 64, 128, 191, 255, 0 };
     VecImage res = add_contrast_omp(image);
-    for (int i = 0; i < image.size(); i++) {
+    for (size_t i = 0; i < image.size(); i++) {
         ASSERT_EQ(correct[i], res[i]);
     }
 }
@@ -54,7 +54,7 @@ TEST(Linear_stretch_contrast_OMP, Seq_vs_Omp) {
     t2 = omp_get_wtime();
     dt_omp = t2 - t1;
 
-    for (int i = 0; i < image.size(); i++) {
+    for (size_t i = 0; i < image.size(); i++) {
         ASSERT_EQ(res_seq[i], res_omp[i]);
     }
